move domain under internal/pkg where it belongs
