+++ internal/pkg/application/facilities/sportsvenues.go
@@ -16,7 +16,7 @@ import (
 	"github.com/diwise/context-broker/pkg/ngsild/types/entities"
 	. "github.com/diwise/context-broker/pkg/ngsild/types/entities/decorators"
 	"github.com/diwise/context-broker/pkg/ngsild/types/properties"
-	"github.com/diwise/integration-cip-sdl/internal/domain"
+	"github.com/diwise/integration-cip-sdl/internal/pkg/domain"
 )
 
 var ErrSportsVenueIsOfIgnoredType error = errors.New("sportsvenue is of non supported type")
